requirements.txt

accelerate
bitsandbytes
chromadb
colbert-ai
datasets
langchain
langchain-community
langchain-openai
llama-index
llama-parse
nltk
ollama
openai
qdrant-client
ragas
ragatouille
sentence-transformers
tokenizers
transformers

faiss-cpu
matplotlib
numpy
onnx
onnxruntime
pandas
scikit-learn
scipy
torch
sympy

fastapi
flask
requests
uvicorn
beautifulsoup4

gitpython
pypdf
pyqt
sqlalchemy

pytest

pylate